refactor(#2154): relocate IccCmdLineUtil.h into IccProfLib
IccConnect reached across the tree for a Tools header:

  IccConnect/IccLibConnect/IccCmmConfig.cpp:79
  #include "../../Tools/CmdLine/IccCmdLineUtil.h" // this should probably
                                                  // move into the ProfLib
                                                  // directory

That was the only cross-tree Tools/CmdLine include anywhere under
IccProfLib/, IccConnect/, IccXML/ or IccJSON/, and the TODO on it is
ChrisCoxArt's own (added in #1683). Library code depending on a tool
directory is what put a second copy of these helpers in play, so the
relocation removes the cause rather than the symptom.

The header is unchanged - git records a 100% rename. Every function in
it is `inline`, so there is no CFILES entry, no new exported symbol and
no ABI change; this is the file-location half of #2154, not the export
half.

Consumers now include it unqualified. IccProfLib's own
target_include_directories already publishes .../IccProfLib PUBLIC for
both the shared and static targets, so the 19 tool translation units and
IccConnect resolve it the same way they already resolve IccUtil.h. It is
added to HEADERS_PUBLIC so installed consumers get it beside its
siblings in include/RefIccMAX/IccProfLib2/.

The warning-policy probe in ci-iccdev-tool-tests.yml compiles this
header by repo-root-relative path, so it follows the move.

Verified: full tree builds clean (tools + IccConnect2 shared/static);
ctest failure set is byte-identical to unmodified master 837dd1ed built
the same way, so nothing regressed; `cmake --install` places the header
in IccProfLib2/ beside IccUtil.h; both the g++ and clang++ legs of the
warning-policy probe pass against the new path; and iccProfilePlot still
writes a raster to a regular file while refusing /dev/null, which is the
S_ISREG guard in the relocated header doing its job.
