+++ Tools/CmdLine/IccApplyNamedCmm/iccApplyNamedCmm.cpp
@@ -78,7 +78,7 @@
 #include "IccProfLibVer.h"
 #include "IccLibConnectVer.h"
 #include "IccConnect.h"
-#include "../IccCmdLineUtil.h"
+#include "IccCmdLineUtil.h"
 #if !defined(_WIN32)
 #include <fcntl.h>
 #include <sys/stat.h>
