+++ Tools/CmdLine/IccApplySearch/iccApplySearch.cpp
@@ -77,7 +77,7 @@
 #include "IccLibConnectVer.h"
 #include "IccSearch.h"
 #include "IccConnect.h"
-#include "../IccCmdLineUtil.h"
+#include "IccCmdLineUtil.h"
 #include <cstdlib>  // EXIT_FAILURE, used by the -cfg argument guard added in #2075
 #include <memory>
 #include <vector>
