+++ Tools/CmdLine/IccProfileVisualize/MiniPDF.cpp
@@ -70,7 +70,7 @@
 #include <cmath>
 #include "vizShared.hpp"
 #include "MiniPDF.hpp"
-#include "../IccCmdLineUtil.h"
+#include "IccCmdLineUtil.h"
 #include "errorLog.hpp"
